Split webserver into separate files
